+++ include/dolphin/os2/OSError.h
