+++ include/dolphin/os2/OSUtil.h
