+++ include/libforest/batconfig.h
@@ -2,8 +2,8 @@
 #define BATCONFIG_H
 
 #include "types.h"
-#include "dolphin/OS/OSMemory.h"
-#include "dolphin/OS/os.h"
+#include "dolphin/os/OSMemory.h"
+#include "dolphin/os.h"
 
 asm void Config24MB();
 asm void Config48MB();
