+++ include/libultra/libultra.h
@@ -2,8 +2,8 @@
 #define LIBULTRA_H
 
 #include "types.h"
-#include "dolphin/OS/OSTime.h"
-#include "dolphin/OS/OSCache.h"
+#include "dolphin/os/OSTime.h"
+#include "dolphin/os/OSCache.h"
 
 int bcmp (void *v1, void *v2, u32 size);
 void bcopy(void *dst, void *src, size_t n);
